% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3Functions.R
\name{call.primer3}
\alias{call.primer3}
\title{call primer3 for a given set of DNAstringSet object}
\usage{
call.primer3(seq, seq_id, product_size_range = c(90, 120), Tm = c(60,
  62.5, 65), primer_size_range = c(15, 18, 21), seq_target = NULL,
  ovl_target = NULL, n = 10)
}
\arguments{
\item{seq}{The DNA string for which primer pairs shall be found.}

\item{seq_id}{The sequence id of the DNA string. Just anything at the moment.}

\item{product_size_range}{default: c(90-120)}

\item{Tm}{melting temprature parameters default: c(60, 62.5, 65)}

\item{primer_size_range}{default: c(15, 18, 21)}

\item{seq_target}{A character string of format
"target,length target,length".}

\item{ovl_target}{A character string of space separated overlap targets.}
}
\description{
call primer3 for a given set of DNAstringSet object
}
\author{
CHK modified Altuna Akalin modified Arnaud Krebs' original function
  (for original function see: https://gist.github.com/al2na/8540391).
}
