% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3Functions.R
\name{format.primer3}
\alias{format.primer3}
\title{Format primer3 Output File}
\usage{
\method{format}{primer3}(primer3_output)
}
\arguments{
\item{primer3_output}{Output file from primer3.}
}
\description{
Format primer3 Output File
}
