% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3Functions.R
\name{get.junctions}
\alias{get.junctions}
\title{Get The Exon Exon Junctions For A Transcript From A GTF}
\usage{
get.junctions(gtf, seq_id)
}
\arguments{
\item{gtf}{Gtf to look in.}

\item{seq_id}{Transcript to look for.}
}
\description{
Get The Exon Exon Junctions For A Transcript From A GTF
}
