% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3Functions.R
\name{get.seq}
\alias{get.seq}
\title{Get A Single Sequence From A Multi Sequence Fasta File}
\usage{
get.seq(fasta, seq_id)
}
\arguments{
\item{fasta}{Fasta to look in for sequence.}

\item{seq_id}{Sequence to look for.}
}
\description{
Get A Single Sequence From A Multi Sequence Fasta File
}
