% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3Functions.R
\name{get.transcript}
\alias{get.transcript}
\title{Get A Transcript Out Of A gtf File}
\usage{
get.transcript(gtf, seq_id)
}
\arguments{
\item{gtf}{Gtf to look in for sequence.}

\item{seq_id}{Sequence to look for.}
}
\description{
Get A Transcript Out Of A gtf File
}
